+++ packages/epo-react-lib/src/atomic/Image/Image.test.tsx
@@ -1,49 +1,79 @@
 import { render, screen } from "@testing-library/react";
-import Image from ".";
+import { composeStory } from "@storybook/react";
+import Meta, { Primary } from "./Image.stories";
+import { stringifySizes, stringifySrcSet } from "./utils";
 
-const props = {
-  image: {
-    altText: "A placeholder image",
-    url: "https://via.placeholder.com/150",
-    url2x: "https://via.placeholder.com/300",
-    url3x: "https://via.placeholder.com/450",
-    width: 150,
-    height: 150,
-  },
-};
+const Image = composeStory(Primary, Meta);
 
-test("should attach url to src", () => {
-  render(<Image {...props} />);
+describe("Image", () => {
+  it("should attach url to src", () => {
+    render(<Image />);
 
-  const img = screen.getByRole("img");
+    const img = screen.getByRole("img");
 
-  const urls = [props.image.url, props.image.url2x, props.image.url3x]
-    .filter((url) => url !== undefined)
-    .map((url, i) => `${url} ${i + 1}x`)
-    .join(", ");
+    expect(img).toHaveAttribute("src", Image.args.image?.url);
+  });
 
-  expect(img).toHaveAttribute("src", expect.stringMatching(props.image.url));
-  expect(img).toHaveAttribute("srcSet", expect.stringMatching(urls));
+  it("should create a srcset", () => {
+    render(<Image />);
+
+    const img = screen.getByRole("img");
+
+    expect(img).toHaveAttribute("srcset");
+  });
+
+  it("should have accessible naming", () => {
+    render(<Image />);
+
+    const img = screen.getByRole("img");
+    expect(img).toHaveAccessibleName(Image.args.image?.altText);
+  });
+
+  it("should have width and height attached", () => {
+    render(<Image />);
+
+    const img = screen.getByRole("img");
+
+    expect(img).toHaveAttribute("width", Image.args.image?.width.toString());
+    expect(img).toHaveAttribute("height", Image.args.image?.height.toString());
+  });
 });
 
-test("should have accessible naming", () => {
-  render(<Image {...props} />);
+describe("stringifySrcSet", () => {
+  it("should transform an array of objects into a srcset string", () => {
+    const srcSet = [
+      { src: "https://rubinobservatory.org", size: 100 },
+      { src: "https://rubinobservatory.org", size: "2x" },
+    ];
 
-  const img = screen.getByRole("img");
-  expect(img).toHaveAccessibleName(props.image.altText);
+    const result = stringifySrcSet(srcSet);
+
+    expect(result).toBe(
+      `${srcSet[0].src} ${srcSet[0].size}w, ${srcSet[1].src} ${srcSet[1].size}`
+    );
+  });
 });
 
-test("should have width and height attached", () => {
-  render(<Image {...props} />);
+const sizes = [
+  { size: "30em" },
+  { size: 101, mediaCondition: "(max-width: 200px)" },
+  { size: 202, mediaCondition: "(max-width: 300px)" },
+];
 
-  const img = screen.getByRole("img");
-
-  expect(img).toHaveAttribute(
-    "width",
-    expect.stringContaining(props.image.width.toString())
-  );
-  expect(img).toHaveAttribute(
-    "height",
-    expect.stringContaining(props.image.height.toString())
-  );
+describe("stringifySizes", () => {
+  it("should move sizes without media conditions to the end", () => {
+    const result = stringifySizes(sizes).split(",");
+    expect(result[0].includes("(")).toBeTruthy();
+    expect(result[result.length - 1].includes("(")).toBeFalsy();
+  });
+  it("should turn numeric sizes to pixels", () => {
+    const result = stringifySizes(sizes);
+    expect(result.includes(`${sizes[1].size}px`)).toBeTruthy();
+  });
+  it("should create a valid sizes string", () => {
+    const result = stringifySizes(sizes);
+    expect(result).toBe(
+      `${sizes[0].mediaCondition} ${sizes[0].size}px, ${sizes[1].mediaCondition} ${sizes[1].size}px, ${sizes[2].size}`
+    );
+  });
 });